added ps and ethtool aware check_mk output.
added ps and ethtool aware check_mk output. Please note that the output
of ps is somewhat limited to the capabilities to the ps command of
busybox for which conversion routines might be a solution in future. But
for the moment this should be enough to actually monitor if a certain
process is running or not.